backend/requirements.txt
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
openai>=1.0.0
chromadb>=0.4.22
pymongo>=4.6.1
python-dotenv>=1.0.0
fastapi>=0.108.0
uvicorn>=0.25.0
pydantic>=2.5.3
sqlalchemy==2.0.23
python-multipart>=0.0.6
Werkzeug==3.0.1
PyPDF2==3.0.1
python-docx==1.0.1
scikit-learn>=1.6.0
numpy>=1.19.5
scipy>=1.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.1
